Refactor Cart and Checkout components for improved structure and readability
- Simplified CartPage by extracting cart logic into a custom hook (useCartPageLogic).
- Replaced inline cart item rendering with CartItemCard component for better modularity.
- Created CheckoutForm and CheckoutOrderSummary components to encapsulate form and summary logic.
- Enhanced error handling and loading states in CheckoutPage.
- Updated payment method handling to use specific types for better type safety.
